Corrigido tipo da variavel nota para float

+++ ex007.py
@@ -1,5 +1,5 @@
 #Desenvolva um programa que leia as duas notas de um aluno, calcule e mostre a sua média
-materia_a = int(input('Digite a nota obtida na Matéria A:'))
-materia_b = int(input('Digite a nota obtida na Matéria B:'))
+materia_a = float(input('Digite a nota obtida na Matéria A:'))
+materia_b = float(input('Digite a nota obtida na Matéria B:'))
 media = (materia_a + materia_b) / 2
 print('A média do aluno foi: {}' .format(media))
